python-ws/requirements.txt
websockets==15.0
requests==2.32.3
python-dotenv==1.1.0
